+++ ixmp/testing.py
@@ -15,14 +15,13 @@ These include:
 
   .. autosummary::
      assert_logs
-     assert_qty_allclose
-     assert_qty_equal
 
 - Methods for setting up and populating test ixmp databases:
 
   .. autosummary::
-     make_dantzig
+     add_test_data
      create_test_platform
+     make_dantzig
      populate_test_platform
 
 - Methods to run and retrieve values from Jupyter notebooks:
@@ -53,14 +52,15 @@ import sys
 
 import numpy as np
 import pandas as pd
+import pint
 import pytest
+import xarray as xr
 from click.testing import CliRunner
 
 from . import cli
 from . import config as ixmp_config
 from .core import IAMC_IDX, Platform, Scenario, TimeSeries
 from .reporting import Quantity
-from .reporting.testing import assert_qty_allclose, assert_qty_equal  # noqa: F401
 
 log = logging.getLogger(__name__)
 
@@ -186,6 +186,37 @@ def bool_param_id(name):
 
 # Create and populate ixmp databases
 
+
+def add_test_data(scen: Scenario):
+    # New sets
+    t_foo = ["foo{}".format(i) for i in (1, 2, 3)]
+    t_bar = ["bar{}".format(i) for i in (4, 5, 6)]
+    t = t_foo + t_bar
+    y = list(map(str, range(2000, 2051, 10)))
+
+    # Add to scenario
+    scen.init_set("t")
+    scen.add_set("t", t)
+    scen.init_set("y")
+    scen.add_set("y", y)
+
+    # Data
+    ureg = pint.get_application_registry()
+    x = Quantity(
+        xr.DataArray(np.random.rand(len(t), len(y)), coords=[("t", t), ("y", y)]),
+        units=ureg.kg,
+    )
+
+    # As a pd.DataFrame with units
+    x_df = x.to_series().rename("value").reset_index()
+    x_df["unit"] = "kg"
+
+    scen.init_par("x", ["t", "y"])
+    scen.add_par("x", x_df)
+
+    return t, t_foo, t_bar, x
+
+
 MODEL = "canning problem"
 SCENARIO = "standard"
 HIST_DF = pd.DataFrame(
@@ -286,7 +317,7 @@ def populate_test_platform(platform):
     s4.set_as_default()
 
 
-def make_dantzig(mp, solve=False):
+def make_dantzig(mp: Platform, solve=False) -> Scenario:
     """Return :class:`ixmp.Scenario` of Dantzig's canning/transport problem.
 
     Parameters