+++ ixmp/tests/reporting/test_computations.py
@@ -1,142 +1,93 @@
 import logging
+from functools import partial
 
-import numpy as np
-import pint
+import pandas as pd
 import pytest
-import xarray as xr
-from pandas.testing import assert_series_equal
+from genno import Computer, Quantity
+from genno.testing import assert_qty_equal
+from pandas.testing import assert_frame_equal
 
-import ixmp
-from ixmp.reporting import Quantity, Reporter, computations
-from ixmp.reporting.testing import random_qty
-from ixmp.testing import assert_logs, assert_qty_equal
-
-from . import add_test_data
+from ixmp.model.dantzig import DATA
+from ixmp.reporting.computations import map_as_qty, update_scenario
+from ixmp.testing import assert_logs, make_dantzig
 
 pytestmark = pytest.mark.usefixtures("parametrize_quantity_class")
 
 
-@pytest.fixture(scope="function")
-def data(test_mp, request):
-    scen = ixmp.Scenario(test_mp, request.node.name, request.node.name, "new")
-    data_objs = list(add_test_data(scen))
-    rep = Reporter.from_scenario(scen)
-    yield [scen, rep] + data_objs
-
-
-@pytest.mark.parametrize(
-    "operands, size",
-    [
-        (("a", "a"), 18),
-        (("a", "x"), 36),
-        (("x", "b"), 36),
-        (("a", "b"), 36),
-        (("a", "x", "b"), 36),
-    ],
-)
-def test_add(data, operands, size):
-    scen, rep, t, t_foo, t_bar, x = data
-
-    y = scen.set("y").tolist()
-    x = rep.get("x:t-y")
-    a = Quantity(
-        xr.DataArray(
-            np.random.rand(len(t_foo), len(y)), coords=[t_foo, y], dims=["t", "y"]
-        ),
-        units=x.attrs["_unit"],
-    )
-    b = Quantity(
-        xr.DataArray(
-            np.random.rand(len(t_bar), len(y)), coords=[t_bar, y], dims=["t", "y"]
-        ),
-        units=x.attrs["_unit"],
+def test_map_as_qty():
+    b = ["b1", "b2", "b3", "b4"]
+    input = pd.DataFrame(
+        [["f1", "b1"], ["f1", "b2"], ["f2", "b3"]], columns=["foo", "bar"]
     )
 
-    rep.add("a:t-y", a)
-    rep.add("b:t-y", b)
+    result = map_as_qty(input, b)
 
-    key = rep.add(
-        "result", tuple([computations.add] + [f"{name}:t-y" for name in operands])
-    )
-
-    result = rep.get(key)
-    assert size == result.size, result.to_series()
-
-
-def test_apply_units(data, caplog):
-    # Unpack
-    *_, x = data
-
-    registry = pint.get_application_registry()
-
-    # Brute-force replacement with incompatible units
-    with assert_logs(caplog, "Replace 'kilogram' with incompatible 'liter'"):
-        result = computations.apply_units(x, "litres")
-    assert result.attrs["_unit"] == registry.Unit("litre")
-    # No change in values
-    assert_series_equal(result.to_series(), x.to_series())
-
-    # Compatible units: magnitudes are also converted
-    with assert_logs(
-        caplog, "Convert 'kilogram' to 'metric_ton'", at_level=logging.DEBUG
-    ):
-        result = computations.apply_units(x, "tonne")
-    assert result.attrs["_unit"] == registry.Unit("tonne")
-    assert_series_equal(result.to_series(), x.to_series() * 0.001)
-
-    # Remove unit
-    x.attrs["_unit"] = registry.Unit("dimensionless")
-
-    caplog.clear()
-    result = computations.apply_units(x, "kg")
-    # Nothing logged when _unit attr is missing
-    assert len(caplog.messages) == 0
-    assert result.attrs["_unit"] == registry.Unit("kg")
-    assert_series_equal(result.to_series(), x.to_series())
-
-
-@pytest.mark.xfail(reason="Outer join of non-intersecting dimensions (AttrSeries only)")
-def test_product0():
-    A = Quantity(xr.DataArray([1, 2], coords=[["a0", "a1"]], dims=["a"]))
-    B = Quantity(xr.DataArray([3, 4], coords=[["b0", "b1"]], dims=["b"]))
     exp = Quantity(
-        xr.DataArray(
-            [[3, 4], [6, 8]],
-            coords=[["a0", "a1"], ["b0", "b1"]],
-            dims=["a", "b"],
-        ),
-        units="1",
+        pd.DataFrame(
+            [
+                ["f1", "b1", 1],
+                ["f1", "b2", 1],
+                ["f2", "b3", 1],
+                ["all", "b1", 1],
+                ["all", "b2", 1],
+                ["all", "b3", 1],
+                ["all", "b4", 1],
+            ],
+            columns=["foo", "bar", "value"],
+        ).set_index(["foo", "bar"])
     )
 
-    assert_qty_equal(exp, computations.product(A, B))
-    computations.product(exp, B)
+    assert_qty_equal(exp, result)
 
 
-def test_product1():
-    """Product of quantities with overlapping dimensions."""
-    A = random_qty(dict(a=2, b=2, c=2, d=2))
-    B = random_qty(dict(b=2, c=2, d=2, e=2, f=2))
+def test_update_scenario(request, caplog, test_mp):
+    scen = make_dantzig(test_mp)
+    scen.check_out()
+    scen.add_set("j", "toronto")
+    scen.commit("Add j=toronto")
 
-    assert computations.product(A, B).size == 2 ** 6
+    # Number of rows in the 'd' parameter
+    N_before = len(scen.par("d"))
+    assert 6 == N_before
 
+    # A Computer used as calculation engine
+    c = Computer()
 
-def test_select(data):
-    # Unpack
-    *_, t_foo, t_bar, x = data
+    # Target Scenario for updating data
+    c.add("target", scen)
 
-    x = Quantity(x)
-    assert x.size == 6 * 6
+    # Create a pd.DataFrame suitable for Scenario.add_par()
+    data = DATA["d"].query("j == 'chicago'").assign(j="toronto")
+    data["value"] += 1.0
 
-    # Selection with inverse=False
-    indexers = {"t": t_foo[0:1] + t_bar[0:1]}
-    result_0 = computations.select(x, indexers=indexers)
-    assert result_0.size == 2 * 6
+    # Add to the Reporter
+    c.add("input", data)
 
-    # Single indexer along one dimension results in 1D data
-    indexers["y"] = "2010"
-    result_1 = computations.select(x, indexers=indexers)
-    assert result_1.size == 2 * 1
+    # Task to update the scenario with the data
+    c.add("test 1", (partial(update_scenario, params=["d"]), "target", "input"))
 
-    # Selection with inverse=True
-    result_2 = computations.select(x, indexers=indexers, inverse=True)
-    assert result_2.size == 4 * 5
+    # Trigger the computation that results in data being added
+    with assert_logs(caplog, f"'d' ← {len(data)} rows", at_level=logging.INFO):
+        # Returns nothing
+        assert c.get("test 1") is None
+
+    # Rows were added to the parameter
+    assert len(scen.par("d")) == N_before + len(data)
+
+    # Modify the data
+    data = pd.concat([DATA["d"], data]).reset_index(drop=True)
+    data["value"] *= 2.0
+
+    # Convert to a Quantity object and re-add
+    q = Quantity(data.set_index(["i", "j"])["value"], name="d", units="km")
+    c.add("input", q)
+
+    # Revise the task; the parameter name ('demand') is read from the Quantity
+    c.add("test 2", (update_scenario, "target", "input"))
+
+    # Trigger the computation
+    with assert_logs(caplog, f"'d' ← {len(data)} rows", at_level=logging.INFO):
+        c.get("test 2")
+
+    # All the rows have been updated
+    assert_frame_equal(scen.par("d"), data)